Expose GetInitializedObject on RuntimeServices in S.Runtime.dll.
It's properly on FormatterServices, but that has some high-layered baggage
and no good place to put it for RTM.
