+++ src/System.Runtime/src/ApiCompatBaseline.txt
@@ -1,0 +1,1 @@
+MembersMustExist : Member 'System.Runtime.CompilerServices.RuntimeHelpers.GetUninitializedObject(System.Type)' does not exist in the implementation but it does exist in the contract.
