+++ src/System.Runtime/tests/System.Runtime.Tests.csproj
@@ -65,7 +65,6 @@
     <Compile Include="System\Runtime\CompilerServices\ConditionalWeakTableTests.cs" />
     <Compile Include="System\Runtime\CompilerServices\StrongBoxTests.cs" />
     <Compile Include="System\Runtime\CompilerServices\RuntimeHelpersTests.cs" />
-    <Compile Include="System\Runtime\Serialization\FormatterServices.cs" />
     <Compile Include="System\SByteTests.cs" />
     <Compile Include="System\SingleTests.cs" />
     <Compile Include="System\StringTests.cs" />
